updating marquee featured component
